--- src/main/java/com/ceos22/cgvclone/domain/movie/Genre.java
@@ -1,19 +1,0 @@
-package com.ceos22.cgvclone.domain.movie;
-
-import jakarta.persistence.*;
-import lombok.*;
-
-@Entity
-@Getter
-@Builder
-@AllArgsConstructor
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class Genre {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @Enumerated(EnumType.STRING)
-    @Builder.Default
-    private GenreType type = GenreType.UNKNOWN;
-}
