--- src/main/java/com/ceos22/cgvclone/domain/movie/GenreType.java
@@ -1,15 +1,0 @@
-package com.ceos22.cgvclone.domain.movie;
-
-public enum GenreType {
-    ACTION,
-    COMEDY,
-    DRAMA,
-    ROMANCE,
-    HORROR,
-    THRILLER,
-    SCI_FI,
-    FANTASY,
-    ANIMATION,
-    DOCUMENTARY,
-    UNKNOWN
-}
