--- src/main/java/com/ceos22/cgvclone/domain/movie/MovieGenre.java
@@ -1,22 +1,0 @@
-package com.ceos22.cgvclone.domain.movie;
-
-import jakarta.persistence.*;
-import lombok.AccessLevel;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Entity
-@Getter
-@IdClass(MovieGenreId.class)
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class MovieGenre {
-    @Id
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "movie_id", nullable = false)
-    private Movie movie;
-
-    @Id
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "genre_id", nullable = false)
-    private Genre genre;
-}
