--- src/main/java/com/ceos22/cgvclone/domain/movie/MovieGenreId.java
@@ -1,13 +1,0 @@
-package com.ceos22.cgvclone.domain.movie;
-
-import lombok.EqualsAndHashCode;
-import lombok.NoArgsConstructor;
-
-import java.io.Serializable;
-
-@NoArgsConstructor
-@EqualsAndHashCode
-public class MovieGenreId implements Serializable {
-    private Long movie;
-    private Long genre;
-}
