--- src/main/java/com/ceos22/cgvclone/domain/movie/MovieRepository.java
@@ -1,6 +1,0 @@
-package com.ceos22.cgvclone.domain.movie;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface MovieRepository extends JpaRepository<Movie, Long> {
-}
