+++ src/main/java/com/ceos22/cgvclone/domain/movie/entity/MovieGenre.java
@@ -1,0 +1,23 @@
+package com.ceos22.cgvclone.domain.movie.entity;
+
+import com.ceos22.cgvclone.domain.movie.key.MovieGenreId;
+import jakarta.persistence.*;
+import lombok.AccessLevel;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Entity
+@Getter
+@IdClass(MovieGenreId.class)
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+public class MovieGenre {
+    @Id
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "movie_id", nullable = false)
+    private Movie movie;
+
+    @Id
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "genre_id", nullable = false)
+    private Genre genre;
+}
