+++ src/main/java/com/ceos22/cgvclone/domain/movie/entity/MovieLike.java
@@ -1,0 +1,30 @@
+package com.ceos22.cgvclone.domain.movie.entity;
+
+import com.ceos22.cgvclone.domain.user.entity.User;
+import jakarta.persistence.*;
+import lombok.AccessLevel;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Entity
+@Getter
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+@Table(
+        name = "movie_likes",
+        uniqueConstraints = {
+                @UniqueConstraint(columnNames = {"user_id", "movie_id"})
+        }
+)
+public class MovieLike {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "movie_id", nullable = false)
+    private Movie movie;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "user_id", nullable = false)
+    private User user;
+}
