+++ src/main/java/com/ceos22/cgvclone/domain/movie/enums/GenreType.java
@@ -1,0 +1,15 @@
+package com.ceos22.cgvclone.domain.movie.enums;
+
+public enum GenreType {
+    ACTION,
+    COMEDY,
+    DRAMA,
+    ROMANCE,
+    HORROR,
+    THRILLER,
+    SCI_FI,
+    FANTASY,
+    ANIMATION,
+    DOCUMENTARY,
+    UNKNOWN
+}
