+++ src/main/java/com/ceos22/cgvclone/domain/movie/repository/MovieRepository.java
@@ -1,0 +1,7 @@
+package com.ceos22.cgvclone.domain.movie.repository;
+
+import com.ceos22.cgvclone.domain.movie.entity.Movie;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface MovieRepository extends JpaRepository<Movie, Long> {
+}
