--- src/main/java/com/ceos22/cgvclone/domain/reservation/Reservation.java
@@ -1,35 +1,0 @@
-package com.ceos22.cgvclone.domain.reservation;
-
-import com.ceos22.cgvclone.domain.common.BaseTimeEntity;
-import com.ceos22.cgvclone.domain.theater.Showtime;
-import com.ceos22.cgvclone.domain.user.User;
-import jakarta.persistence.*;
-import lombok.*;
-import org.hibernate.annotations.CreationTimestamp;
-import org.hibernate.annotations.UuidGenerator;
-
-import java.time.LocalDateTime;
-import java.util.UUID;
-
-@Entity
-@Getter
-@Builder
-@AllArgsConstructor
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class Reservation extends BaseTimeEntity {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @UuidGenerator
-    @Column(columnDefinition = "BINARY(16)", updatable = false, nullable = false)
-    private UUID uuid;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "user_id", nullable = false)
-    private User User;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(nullable = false)
-    private Showtime showtime;
-}
