--- src/main/java/com/ceos22/cgvclone/domain/reservation/ReservationRepository.java
@@ -1,6 +1,0 @@
-package com.ceos22.cgvclone.domain.reservation;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface ReservationRepository extends JpaRepository<Reservation, Long> {
-}
