--- src/main/java/com/ceos22/cgvclone/domain/reservation/ReservationSeat.java
@@ -1,25 +1,0 @@
-package com.ceos22.cgvclone.domain.reservation;
-
-import com.ceos22.cgvclone.domain.theater.Seat;
-import jakarta.persistence.*;
-import lombok.AccessLevel;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Entity
-@Getter
-@IdClass(ReservationSeatId.class)
-@AllArgsConstructor
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class ReservationSeat {
-    @Id
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "reservation_id", nullable = false)
-    private Reservation reservation;
-
-    @Id
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "seat_id", nullable = false)
-    private Seat seat;
-}
