--- src/main/java/com/ceos22/cgvclone/domain/reservation/ReservationSeatId.java
@@ -1,13 +1,0 @@
-package com.ceos22.cgvclone.domain.reservation;
-
-import lombok.EqualsAndHashCode;
-import lombok.NoArgsConstructor;
-
-import java.io.Serializable;
-
-@NoArgsConstructor
-@EqualsAndHashCode
-public class ReservationSeatId implements Serializable {
-    private Long reservation;
-    private Long seat;
-}
