+++ src/main/java/com/ceos22/cgvclone/domain/reservation/entity/Reservation.java
@@ -1,0 +1,33 @@
+package com.ceos22.cgvclone.domain.reservation.entity;
+
+import com.ceos22.cgvclone.domain.common.BaseTimeEntity;
+import com.ceos22.cgvclone.domain.theater.entity.Showtime;
+import com.ceos22.cgvclone.domain.user.entity.User;
+import jakarta.persistence.*;
+import lombok.*;
+import org.hibernate.annotations.UuidGenerator;
+
+import java.util.UUID;
+
+@Entity
+@Getter
+@Builder
+@AllArgsConstructor
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+public class Reservation extends BaseTimeEntity {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @UuidGenerator
+    @Column(columnDefinition = "BINARY(16)", updatable = false, nullable = false)
+    private UUID uuid;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "user_id", nullable = false)
+    private User User;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(nullable = false)
+    private Showtime showtime;
+}
