+++ src/main/java/com/ceos22/cgvclone/domain/reservation/entity/ReservationSeat.java
@@ -1,0 +1,26 @@
+package com.ceos22.cgvclone.domain.reservation.entity;
+
+import com.ceos22.cgvclone.domain.reservation.key.ReservationSeatId;
+import com.ceos22.cgvclone.domain.theater.entity.Seat;
+import jakarta.persistence.*;
+import lombok.AccessLevel;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Entity
+@Getter
+@IdClass(ReservationSeatId.class)
+@AllArgsConstructor
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+public class ReservationSeat {
+    @Id
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "reservation_id", nullable = false)
+    private Reservation reservation;
+
+    @Id
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "seat_id", nullable = false)
+    private Seat seat;
+}
