--- src/main/java/com/ceos22/cgvclone/domain/snack/InventoryRepository.java
@@ -1,6 +1,0 @@
-package com.ceos22.cgvclone.domain.snack;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface InventoryRepository extends JpaRepository<Inventory, Integer> {
-}
