--- src/main/java/com/ceos22/cgvclone/domain/snack/OrderItem.java
@@ -1,33 +1,0 @@
-package com.ceos22.cgvclone.domain.snack;
-
-import jakarta.persistence.*;
-import lombok.AccessLevel;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Entity
-@Getter
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@Table(
-        name = "order_items",
-        uniqueConstraints = {
-                @UniqueConstraint(columnNames = {"order_id", "product_id"})
-        }
-)
-public class OrderItem {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "order_id", nullable = false)
-    private UserOrder order;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "product_id", nullable = false)
-    private Product product;
-
-    private int quantity;
-
-    private int price;  // 주문 시점의 가격
-}
