--- src/main/java/com/ceos22/cgvclone/domain/snack/OrderRepository.java
@@ -1,6 +1,0 @@
-package com.ceos22.cgvclone.domain.snack;
-
-import org.springframework.data.repository.CrudRepository;
-
-public interface OrderRepository extends CrudRepository<UserOrder, Long> {
-}
