--- src/main/java/com/ceos22/cgvclone/domain/snack/Product.java
@@ -1,24 +1,0 @@
-package com.ceos22.cgvclone.domain.snack;
-
-import jakarta.persistence.*;
-import lombok.*;
-
-@Entity
-@Getter
-@Builder
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@AllArgsConstructor
-public class Product {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @Column(nullable = false)
-    private String name;
-
-    @Column(nullable = false)
-    private String vendor;
-
-    @Column(nullable = false)
-    private String price;
-}
