--- src/main/java/com/ceos22/cgvclone/domain/snack/ProductRepository.java
@@ -1,6 +1,0 @@
-package com.ceos22.cgvclone.domain.snack;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface ProductRepository extends JpaRepository<Product, Long> {
-}
