--- src/main/java/com/ceos22/cgvclone/domain/snack/UserOrder.java
@@ -1,36 +1,0 @@
-package com.ceos22.cgvclone.domain.snack;
-
-import com.ceos22.cgvclone.domain.common.BaseTimeEntity;
-import com.ceos22.cgvclone.domain.theater.Theater;
-import com.ceos22.cgvclone.domain.user.User;
-import jakarta.persistence.*;
-import lombok.*;
-import org.hibernate.annotations.UuidGenerator;
-
-import java.util.UUID;
-
-@Entity
-@Getter
-@Builder
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@AllArgsConstructor
-public class UserOrder extends BaseTimeEntity {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @UuidGenerator
-    @Column(columnDefinition = "BINARY(16)", updatable = false, nullable = false)
-    private UUID uuid;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "user_id", nullable = false)
-    private User user;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "theater_id", nullable = false)
-    private Theater theater;
-
-    @Builder.Default
-    private Long total_price = 0L;
-}
