+++ src/main/java/com/ceos22/cgvclone/domain/snack/entity/Product.java
@@ -1,0 +1,24 @@
+package com.ceos22.cgvclone.domain.snack.entity;
+
+import jakarta.persistence.*;
+import lombok.*;
+
+@Entity
+@Getter
+@Builder
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+@AllArgsConstructor
+public class Product {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @Column(nullable = false)
+    private String name;
+
+    @Column(nullable = false)
+    private String vendor;
+
+    @Column(nullable = false)
+    private String price;
+}
