--- src/main/java/com/ceos22/cgvclone/domain/theater/Screen.java
@@ -1,25 +1,0 @@
-package com.ceos22.cgvclone.domain.theater;
-
-import jakarta.persistence.*;
-import lombok.*;
-
-@Entity
-@Getter
-@Builder
-@AllArgsConstructor
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class Screen {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "theater_id")
-    private Theater theater;
-
-    @Enumerated(EnumType.STRING)
-    @Builder.Default
-    private ScreenType type = ScreenType.NORMAL;
-
-    private String name;
-}
