--- src/main/java/com/ceos22/cgvclone/domain/theater/Seat.java
@@ -1,23 +1,0 @@
-package com.ceos22.cgvclone.domain.theater;
-
-import jakarta.persistence.*;
-import lombok.*;
-
-@Entity
-@Getter
-@Builder
-@AllArgsConstructor
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class Seat {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "screen_id", nullable = false)
-    private Screen screen;
-
-    private String rowNo;
-
-    private int colNo;
-}
