--- src/main/java/com/ceos22/cgvclone/domain/theater/Theater.java
@@ -1,27 +1,0 @@
-package com.ceos22.cgvclone.domain.theater;
-
-import jakarta.persistence.*;
-import lombok.*;
-import org.hibernate.annotations.Fetch;
-
-import java.util.ArrayList;
-import java.util.List;
-
-@Entity
-@Getter
-@Builder
-@AllArgsConstructor
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class Theater {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    private String name;
-
-    private String location;
-
-    @OneToMany(mappedBy = "theater", fetch = FetchType.LAZY)
-    private List<Screen> screens = new ArrayList<>();
-}
