--- src/main/java/com/ceos22/cgvclone/domain/theater/TheaterLike.java
@@ -1,30 +1,0 @@
-package com.ceos22.cgvclone.domain.theater;
-
-import com.ceos22.cgvclone.domain.user.User;
-import jakarta.persistence.*;
-import lombok.AccessLevel;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Entity
-@Getter
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@Table(
-        name = "theater_likes",
-        uniqueConstraints = {
-                @UniqueConstraint(columnNames = {"theater_id", "user_id"})
-        }
-)
-public class TheaterLike {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "theater_id", nullable = false)
-    private Theater theater;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "user_id", nullable = false)
-    private User user;
-}
