--- src/main/java/com/ceos22/cgvclone/domain/theater/TheaterLikeRepository.java
@@ -1,6 +1,0 @@
-package com.ceos22.cgvclone.domain.theater;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface TheaterLikeRepository extends JpaRepository<TheaterLike, Long> {
-}
