+++ src/main/java/com/ceos22/cgvclone/domain/theater/entity/Showtime.java
@@ -1,0 +1,30 @@
+package com.ceos22.cgvclone.domain.theater.entity;
+
+import com.ceos22.cgvclone.domain.movie.entity.Movie;
+import jakarta.persistence.*;
+import lombok.*;
+
+
+import java.time.LocalDateTime;
+
+@Entity
+@Getter
+@Builder
+@AllArgsConstructor
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+public class Showtime {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "movie_id", nullable = false)
+    private Movie movie;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "screen_id", nullable = false)
+    private Screen screen;
+
+    private LocalDateTime startTime;
+    private LocalDateTime endTime;
+}
