+++ src/main/java/com/ceos22/cgvclone/domain/theater/repository/TheaterLikeRepository.java
@@ -1,0 +1,7 @@
+package com.ceos22.cgvclone.domain.theater.repository;
+
+import com.ceos22.cgvclone.domain.theater.entity.TheaterLike;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface TheaterLikeRepository extends JpaRepository<TheaterLike, Long> {
+}
