+++ src/main/java/com/ceos22/cgvclone/domain/theater/repository/TheaterRepository.java
@@ -1,0 +1,7 @@
+package com.ceos22.cgvclone.domain.theater.repository;
+
+import com.ceos22.cgvclone.domain.theater.entity.Theater;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface TheaterRepository extends JpaRepository<Theater, Long> {
+}
