--- src/main/java/com/ceos22/cgvclone/domain/user/UserRepository.java
@@ -1,5 +1,0 @@
-package com.ceos22.cgvclone.domain.user;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface UserRepository extends JpaRepository<User, Long> {}
