+++ src/main/java/com/ceos22/cgvclone/domain/user/entity/User.java
@@ -1,0 +1,35 @@
+package com.ceos22.cgvclone.domain.user.entity;
+
+import com.ceos22.cgvclone.domain.common.BaseTimeEntity;
+import jakarta.persistence.*;
+import lombok.*;
+import org.hibernate.annotations.UuidGenerator;
+
+import java.util.UUID;
+
+@Table(name = "users")
+@Entity
+@Getter
+@Builder
+@AllArgsConstructor
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+public class User extends BaseTimeEntity {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @UuidGenerator
+    @Column(columnDefinition = "BINARY(16)", updatable = false, nullable = false)
+    private UUID uuid;
+
+    @Column(unique = true, nullable = false)
+    private String email;
+
+    @Column(nullable = false)
+    private String password;
+
+    @Column(nullable = false)
+    private String name;
+
+    private String phoneNumber;
+}
