+++ src/main/java/com/ceos22/cgvclone/domain/user/repository/UserRepository.java
@@ -1,0 +1,6 @@
+package com.ceos22.cgvclone.domain.user.repository;
+
+import com.ceos22.cgvclone.domain.user.entity.User;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface UserRepository extends JpaRepository<User, Long> {}
